requirements.txt
fastapi
uvicorn
reportlab
pydantic
python-dotenv
google-generativeai